pyproject.toml
[tool.poetry]
name = "ubdc-airbnb"
version = "0.1.0"
description = ""
authors = ["nikos ves <nick.ves@glasgow.ac.uk>"]
readme = "README.md"
packages = [
    { include = "ubdc_airbnb", from = 'src/ubdc_airbnb'},
    { include = "core", from = 'src/ubdc_airbnb'}
]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4"
celery = '^5'
kombu = { version = "v5.3.0b2", allow-prereleases = true }
django-celery-beat = "~2.4.0"
django-celery-results = "~2.4.0"
jsonpath-ng = "^1.5.3"
more-itertools = "^9.0.0"
psycopg2-binary = "^2"
mercantile = "^1.2.1"
python-dotenv = "^0.21.0"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.2.1"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
requests-mock = "^1.10.0"
pytest-env = "^0.8.1"
faker = "^16.6.1"
pytest-celery = "^0.0.0"
responses = "^0.22.0"
pylint = "^2.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
env = [
    "DJANGO_ENV=TEST"
]
addopts = [
    "--import-mode=importlib",
]
DJANGO_SETTINGS_MODULE = "core.settings"
